// src/service/summoner/model/get-league-response-data-model.ts
export default interface getLeagueResponseDataModel {
  freshBlood: boolean
  hotStreak: boolean
  inactive: boolean
  leagueId: string
  leaguePoints: number
  losses: number
  queueType: string
  rank: string
  summonerId: string
  summonerName: string
  tier: string
  veteran: boolean
  wins: number
}